+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## Unreleased changes
+
+## v0.3.2
+
+* Add `date` command.
+* Add `lsblk` and `blkread` commands.
+* Renamed `bioshw` to `lshw`
+
 ## v0.3.1
 
 * Add `hexdump`, `load` and `run` commands.